+++ lib/st_steps/st_steps.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 
-import 'package:saturn/st_icons/st_icons.dart';
 import 'package:saturn/st_steps/common.dart';
+import 'package:saturn/st_icons/st_icons.dart';
+import 'package:saturn/st_steps/steps_stateful.dart';
 
 const _defaultCircleWidth = 8.0;
 const _defaultNumIconWidth = 20.0;
-const _defaultNumIconHeight = 22.0;
 const _defaultMargin = EdgeInsets.symmetric(horizontal: 16.0);
 const _defaultTextStyle = TextStyle(color: Color(0xFF555555), fontSize: 14);
 const _defaultSelectTextStyle =
@@ -14,16 +14,17 @@ const _defaultNumTextStyle = TextStyle(color: Color(0xFFFFFFFF), fontSize: 14);
 const _defaultColor = Color(0xFFDFE2E7);
 const _defaultSelectColor = Color(0xFF095BF9);
 
-class STSteps extends StatefulWidget {
-  const STSteps({
-    Key key,
-    this.type = STStepsType.dot,
-    this.margin,
-    this.steps,
-    this.current,
-    this.deatilHeight,
-    this.detailWidth,
-  }) : super(key: key);
+class STSteps extends StatelessWidget {
+  const STSteps(
+      {Key key,
+      this.type = STStepsType.dot,
+      this.margin = _defaultMargin,
+      this.steps,
+      this.current = 1,
+      this.deatilHeight,
+      this.detailWidth})
+      : assert(steps.length > 1),
+        super(key: key);
 
   final STStepsType type;
   final EdgeInsets margin;
@@ -33,54 +34,8 @@ class STSteps extends StatefulWidget {
   final double detailWidth; // type为detail,竖排需固定宽度才能满足外部的对齐方式
 
   @override
-  _STStepsState createState() => _STStepsState();
-}
-
-class _STStepsState extends State<STSteps> {
-  STStepsType _type;
-  List<STStepItem> _steps;
-  EdgeInsets _margin;
-  List<GlobalKey> _golbalKeys;
-  bool _isFindRender = false;
-  List<STRenderItem> _renders;
-  int _current;
-  double _detailHeight;
-  double _detailWidth;
-
-  @override
-  void initState() {
-    super.initState();
-    _type = widget.type;
-    _steps = List.from(widget.steps);
-    _margin = widget.margin ?? _defaultMargin;
-    _current = widget.current ?? 1;
-    _golbalKeys = <GlobalKey>[];
-    for (var i = 0; i < _steps.length; i++) {
-      final _tempKey = GlobalKey(debugLabel: '${_steps[i].title}+$i');
-      _golbalKeys.add(_tempKey);
-    }
-
-    WidgetsBinding.instance.addPostFrameCallback((Duration timeStamp) {
-      _findRenderObject();
-    });
-  }
-
-  void _findRenderObject() {
-    _isFindRender = true;
-    _renders = <STRenderItem>[];
-    for (var i = 0; i < _golbalKeys.length; i++) {
-      final RenderBox tempRender =
-          _golbalKeys[i].currentContext.findRenderObject();
-      final _offset = tempRender.localToGlobal(Offset.zero);
-      final _render = STRenderItem(offset: _offset, size: tempRender.size);
-      _renders.add(_render);
-    }
-    setState(() {});
-  }
-
-  @override
   Widget build(BuildContext context) {
-    switch (_type) {
+    switch (type) {
       case STStepsType.dot:
         return buildDot();
       case STStepsType.number:
@@ -92,241 +47,92 @@ class _STStepsState extends State<STSteps> {
     return Container();
   }
 
+  Widget buildDot() {
+    return STStatefulSteps(
+      margin: margin,
+      steps: steps,
+      current: current,
+    );
+  }
+
   Widget buildDetail() {
-    _detailHeight = widget.deatilHeight ??
-        (44.0 * _steps.length + 30 * (_steps.length - 1));
-    _detailWidth = widget.detailWidth ?? 100.0;
-    final _stackWidgets = <Widget>[];
-    final _bgWidgtets = List.generate(_steps.length, (index) {
-      Widget _preWidget;
-      if (_isFinished(index)) {
-        _preWidget = const Icon(
-          STIcons.commonly_selected,
-          size: _defaultNumIconWidth * 2 / 3,
-          color: Colors.white,
-        );
-      } else {
-        final _number = _steps[index].number ?? index + 1;
-        _preWidget = Text('$_number', style: _defaultNumTextStyle);
-      }
-      return SizedBox(
-        width: _detailWidth,
-        child: Row(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Container(
-              key: _golbalKeys[index],
-              height: _defaultNumIconWidth,
-              width: _defaultNumIconWidth,
-              decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(_defaultNumIconWidth / 2),
-                color: _isFinished(index) ? _defaultSelectColor : _defaultColor,
-              ),
-              alignment: Alignment.center,
-              child: _preWidget,
-            ),
-            const SizedBox(width: 4),
-            Column(
-              children: [
-                SizedBox(
-                  height: _defaultNumIconHeight,
-                  child: Text(
-                    _steps[index].title,
-                    style: TextStyle(
-                        fontSize: 14.0,
-                        color: _isFinished(index)
-                            ? const Color(0xFF000000)
-                            : const Color(0xFF555555)),
-                  ),
-                ),
-                SizedBox(
-                  height: _defaultNumIconWidth,
-                  child: Text(
-                    _steps[index].info,
-                    style: const TextStyle(
-                        fontSize: 12.0, color: Color(0xFF888888)),
-                  ),
-                ),
-              ],
-            ),
-          ],
-        ),
-      );
-    }).toList();
-    _stackWidgets.add(SizedBox(
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: _bgWidgtets,
-      ),
-    ));
-    if (_isFindRender) {
-      for (var i = 0; i < _steps.length - 1; i++) {
-        final _top = _renders[i].offset.dy -
-            _renders[0].offset.dy +
-            _defaultNumIconWidth +
-            _defaultCircleWidth / 2;
-        final _height = _renders[i + 1].offset.dy -
-            _renders[i].offset.dy -
-            _defaultNumIconWidth -
-            _defaultCircleWidth;
-        final _temp = Positioned(
-          left: (_defaultNumIconWidth - 2) / 2,
-          top: _top,
-          child: Container(
-            height: _height,
-            width: 2,
-            color: _isFinished(i + 1) ? _defaultSelectColor : _defaultColor,
-          ),
-        );
-        _stackWidgets.add(_temp);
-      }
-    }
-    return SizedBox(
-      height: _detailHeight,
-      child: Stack(
-        children: _stackWidgets,
-      ),
+    return STStatefulSteps(
+      type: STStepsType.detail,
+      current: current,
+      margin: margin,
+      steps: steps,
+      detailWidth: detailWidth,
+      deatilHeight: deatilHeight,
     );
   }
 
   Widget buildNumberAndIcon() {
-    final _stackWidgets = <Widget>[];
-    final _bgWidgtets = List.generate(_steps.length, (index) {
-      Widget _preWidget;
-      if (_type == STStepsType.number) {
-        final _number = _steps[index].number ?? index + 1;
-        _preWidget = Text('$_number', style: _defaultNumTextStyle);
-      } else {
-        _preWidget = Icon(
-          _steps[index].iconData ?? STIcons.commonly_selected,
-          size: _defaultNumIconWidth * 2 / 3,
-          color: Colors.white,
-        );
-      }
-      return Row(
-        key: _golbalKeys[index],
-        children: [
-          Container(
-            height: _defaultNumIconWidth,
-            width: _defaultNumIconWidth,
-            decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(_defaultNumIconWidth / 2),
-              color: _isFinished(index) ? _defaultSelectColor : _defaultColor,
-            ),
-            alignment: Alignment.center,
-            child: _preWidget,
-          ),
-          const SizedBox(width: 4),
-          Text(
-            _steps[index].title,
-            style: _isFinished(index)
-                ? _defaultSelectTextStyle
-                : _defaultTextStyle,
-          ),
-        ],
-      );
-    }).toList();
-    _stackWidgets.add(SizedBox(
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: _bgWidgtets,
-      ),
-    ));
-    if (_isFindRender) {
-      for (var i = 1; i < _steps.length; i++) {
-        final _width = _renders[i].offset.dx -
-            _renders[i - 1].offset.dx -
-            _renders[i - 1].size.width -
-            _defaultNumIconWidth / 2;
-        final _left = _renders[i - 1].offset.dx +
-            _renders[i - 1].size.width -
-            _defaultNumIconWidth / 2;
-        final _temp = Positioned(
-          left: _left,
-          bottom: (_defaultNumIconHeight - 2) / 2,
-          child: Container(
-            height: 2,
-            width: _width,
-            color: _isFinished(i) ? _defaultSelectColor : _defaultColor,
-          ),
-        );
-        _stackWidgets.add(_temp);
-      }
-    }
-    return Container(
-      margin: _margin,
-      height: _defaultNumIconHeight,
-      child: Stack(
-        children: _stackWidgets,
-      ),
-    );
-  }
-
-  Widget buildDot() {
-    final _stackWidgets = <Widget>[];
-    final _titlesWidgets = List.generate(
-      _steps.length,
-      (index) {
-        var _crossAxisAlignment = CrossAxisAlignment.center;
-        if (index == 0) {
-          _crossAxisAlignment = CrossAxisAlignment.start;
-        } else if (index == _steps.length - 1) {
-          _crossAxisAlignment = CrossAxisAlignment.end;
+    final _widgets = <Widget>[];
+    for (var i = 0; i < steps.length * 2 - 1; i++) {
+      if (i % 2 == 0) {
+        final j = i ~/ 2;
+        Widget _preWidget;
+        if (type == STStepsType.number) {
+          final _number = steps[j].number ?? j + 1;
+          _preWidget = Text('$_number', style: _defaultNumTextStyle);
+        } else {
+          _preWidget = Icon(
+            steps[j].iconData ?? STIcons.commonly_selected,
+            size: _defaultNumIconWidth * 2 / 3,
+            color: Colors.white,
+          );
         }
-        return Column(
-          crossAxisAlignment: _crossAxisAlignment,
-          children: [
-            Text(
-              _steps[index].title,
-              style: _isFinished(index)
-                  ? _defaultSelectTextStyle
-                  : _defaultTextStyle,
-            ),
-            const SizedBox(height: _defaultCircleWidth / 2),
-            Container(
-              key: _golbalKeys[index],
-              height: _defaultCircleWidth,
-              width: _defaultCircleWidth,
-              decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(_defaultCircleWidth / 2),
-                color: _isFinished(index) ? _defaultSelectColor : _defaultColor,
+        var _padding =
+            const EdgeInsets.symmetric(horizontal: _defaultCircleWidth / 2);
+        if (j == 0) {
+          _padding = const EdgeInsets.only(right: _defaultCircleWidth / 2);
+        } else if (j == steps.length - 1) {
+          _padding = const EdgeInsets.only(left: _defaultCircleWidth / 2);
+        }
+        final _numberIcon = Container(
+          padding: _padding,
+          child: Row(
+            children: [
+              Container(
+                height: _defaultNumIconWidth,
+                width: _defaultNumIconWidth,
+                decoration: BoxDecoration(
+                  borderRadius: BorderRadius.circular(_defaultNumIconWidth / 2),
+                  color: _isFinished(j) ? _defaultSelectColor : _defaultColor,
+                ),
+                alignment: Alignment.center,
+                child: _preWidget,
               ),
-            ),
-          ],
-        );
-      },
-    ).toList();
-    _stackWidgets.add(SizedBox(
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: _titlesWidgets,
-      ),
-    ));
-    if (_isFindRender) {
-      for (var i = 1; i < _steps.length; i++) {
-        final _width = _renders[i].offset.dx -
-            _renders[i - 1].offset.dx -
-            _defaultCircleWidth * 2;
-        final _left = _renders[i - 1].offset.dx - _defaultCircleWidth / 2;
-        final _temp = Positioned(
-          left: _left,
-          bottom: (_defaultCircleWidth - 2) / 2,
-          child: Container(
-            height: 2,
-            width: _width,
-            color: _isFinished(i) ? _defaultSelectColor : _defaultColor,
+              const SizedBox(width: 4),
+              Text(
+                steps[j].title,
+                style: _isFinished(j)
+                    ? _defaultSelectTextStyle
+                    : _defaultTextStyle,
+              ),
+            ],
           ),
         );
-        _stackWidgets.add(_temp);
+        _widgets.add(_numberIcon);
+      } else {
+        final _index = i ~/ 2 + 1;
+        final _flex = Flexible(
+          child: Container(
+            height: 2,
+            color: _isFinished(_index) ? _defaultSelectColor : _defaultColor,
+          ),
+        );
+        _widgets.add(_flex);
       }
     }
+
     return Container(
-      margin: _margin,
-      child: Stack(
-        children: _stackWidgets,
+      margin: margin,
+      child: Row(
+        children: _widgets,
       ),
     );
   }
 
-  bool _isFinished(int index) => index < _current;
+  bool _isFinished(int index) => index < current;
 }
